# docs/swagger.yaml
definitions:
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      timestamp:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.ShortenRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  dto.ShortenResponse:
    properties:
      short_id:
        type: string
      short_url:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticates the user and generates a JWT token.
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to generate token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: User Login
      tags:
      - Authentication
  /api/shorten:
    post:
      consumes:
      - application/json
      description: Receives a long URL and returns a shortened version.
      parameters:
      - description: URL
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/dto.ShortenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Response with shorter URL
          schema:
            $ref: '#/definitions/dto.ShortenResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to shorten URL
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Shorten a URL
      tags:
      - URLs
swagger: "2.0"